feat(toolbar): calendar requested: #210, #160
